% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GED.R
\docType{package}
\name{ucdp}
\alias{ucdp}
\alias{ucdp-package}
\title{Disaggregated Data from the UCDP}
\description{
This package contains the most disaggregated data available
from the Uppsala Conflict Data Program, i.e. the Georeferenced
Event Dataset (GED) Global version 20.1.
}
